requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.2=py39h06a4308_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=unix_pyh707e725_0
cryptography=39.0.1=py39h9ce1e76_2
cudatoolkit=11.3.1=h2bc3f7f_2
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pyh1a96a4e_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_1
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.48.2=py39he1ff14a_1
idna=3.4=py39h06a4308_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46305
joblib=1.3.1=pypi_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pyhd8ed1ab_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
multidict=6.0.2=py39h5eee18b_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.25.0=py39h5f9d8c6_0
numpy-base=1.25.0=py39hb5e798b_0
oauthlib=3.2.2=pyhd8ed1ab_0
opencv-python=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.9=h7f8727e_0
packaging=23.1=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.1.2=py39h06a4308_0
protobuf=4.23.4=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.0.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.17=h955ad1f_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.29.0=py39h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post5=pypi_0
sklearn3=0.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py39h52d8a92_0
tensorboardx=2.6.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.11.0=py39_cu113
torchinfo=1.8.0=pypi_0
torchvision=0.12.0=py39_cu113
tqdm=4.65.0=pypi_0
typing_extensions=4.6.3=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39h06a4308_0
vit-pytorch=1.4.5=pypi_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.7.2=py39hb9d737c_2
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
